refactor(cli): split ProjectGenerator into one partial file per template
ProjectGenerator.cs had grown to 2471 lines with every template's content inlined.
Split it one-partial-per-template (.Server/.Wasm/.Native, plus .Shared.cs for the
content more than one emits), and fold the three identical copies of the
namespace-token replacement pass into a shared Materialize helper.

Pure move: no behaviour change, 241 CLI tests unchanged and green.
